WFLY-15838 Move tests specifically about the Lucene backend to a dedicated package
