+++ testsuite/preview/basic/src/test/java/org/jboss/as/test/integration/hibernate/search/backend/lucene/extension/HibernateSearchLuceneDependencyTestCase.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.hibernate.search.extension;
+package org.jboss.as.test.integration.hibernate.search.backend.lucene.extension;
 
 import org.apache.lucene.index.Term;
 import org.apache.lucene.search.TermQuery;
