+++ testsuite/preview/basic/src/test/java/org/jboss/as/test/integration/hibernate/search/backend/lucene/massindexer/Singer.java
@@ -1,6 +1,6 @@
 /*
  * JBoss, Home of Professional Open Source.
- * Copyright 2018, Red Hat, Inc., and individual contributors
+ * Copyright 2012, Red Hat, Inc., and individual contributors
  * as indicated by the @author tags. See the copyright.txt file in the
  * distribution for a full listing of individual contributors.
  *
@@ -19,27 +19,61 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.integration.hibernate.search.simple;
+package org.jboss.as.test.integration.hibernate.search.backend.lucene.massindexer;
 
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.FullTextField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import java.io.Serializable;
 
 @Entity
 @Indexed
-public class Book {
+public class Singer implements Serializable {
+    private static final long serialVersionUID = 1L;
 
     @Id
-    @GeneratedValue
-    Long id;
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
 
     @FullTextField
-    @FullTextField(name = "title_autocomplete",
-            analyzer = AnalysisConfigurer.AUTOCOMPLETE,
-            searchAnalyzer = AnalysisConfigurer.AUTOCOMPLETE_QUERY)
-    String title;
+    private String firstName;
 
+    @FullTextField
+    private String lastName;
+
+    public long getId() {
+        return this.id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getFirstName() {
+        return this.firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return this.lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    @Override
+    public String toString() {
+        return "[#" + id + "] " + lastName + ", " + firstName;
+    }
 }
+
+
+
