+++ testsuite/preview/basic/src/test/java/org/jboss/as/test/integration/hibernate/search/backend/lucene/simple/persistence.xml
@@ -16,7 +16,7 @@
             <property name="hibernate.search.backend.lucene_version" value="LUCENE_CURRENT"/>
             <property name="hibernate.search.backend.directory.type" value="local-heap"/>
             <property name="hibernate.search.backend.analysis.configurer"
-                      value="org.jboss.as.test.integration.hibernate.search.simple.AnalysisConfigurer"/>
+                      value="org.jboss.as.test.integration.hibernate.search.backend.lucene.simple.AnalysisConfigurer"/>
         </properties>
     </persistence-unit>
 </persistence>
