+++ testsuite/preview/basic/src/test/java/org/jboss/as/test/integration/hibernate/search/v5migrationhelper/massindexer/HibernateSearchV5MigrationHelperMassIndexerTestCase.java
@@ -23,6 +23,7 @@ package org.jboss.as.test.integration.hibernate.search.v5migrationhelper.massind
 
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.as.test.integration.hibernate.search.backend.lucene.massindexer.HibernateSearchLuceneEarMassIndexerTestCase;
 import org.jboss.as.test.shared.util.AssumeTestGroupUtil;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.asset.Asset;
@@ -48,9 +49,9 @@ import static org.junit.Assert.assertTrue;
  * Verify deployed applications can use the mass indexer of the default Hibernate Search module
  * through the V5 migration helper provided as an application library.
  * <p>
- * This test is similar to {@link org.jboss.as.test.integration.hibernate.search.massindexer.HibernateSearchEarMassIndexerTestCase},
+ * This test is similar to {@link HibernateSearchLuceneEarMassIndexerTestCase},
  * but uses old APIs thanks to the V5 migration helper JARs that are packaged with the application.
- * Those are not provided as modules because they shouldn't be used in production.
+ * Those are not provided as WildFly modules because they shouldn't be used in production.
  */
 @RunWith(Arquillian.class)
 public class HibernateSearchV5MigrationHelperMassIndexerTestCase {
